+++ src/clone.rs
@@ -2,9 +2,9 @@
 //    by Lut99
 //
 //  Created:
-//    09 Jan 2025, 01:09:44
-//  Last edited:
 //    04 Feb 2025, 15:41:18
+//  Last edited:
+//    04 Feb 2025, 16:33:51
 //  Auto updated?
 //    Yes
 //